Claude/assist/family-bot/family-bot/requirements.txt
# Install with: pip install -r requirements.txt
python-telegram-bot==20.7
openai>=1.12.0
python-dateutil>=2.8.2
APScheduler>=3.10.4
python-dotenv>=1.0.0
pytz>=2024.1

# python-dateutil already includes rrule — no separate package needed
